+++ CHANGES.rst
@@ -4,7 +4,7 @@
 ramp_fitting
 ------------
 
-- For slopes with negative median rates, the Poisson variance is zero.
+- For slopes with negative median rates, the Poisson variance is zero. [#33]
 
 0.4.1 (2021-10-14)
 ==================