+++ CHANGES.rst
@@ -1,10 +1,11 @@
-(unreleased)
+0.4.2 (2021-10-28)
 ==================
 
 ramp_fitting
 ------------
 
 - For slopes with negative median rates, the Poisson variance is zero. [#59]
+
 - Changed the way the final DQ array gets computed when handling the DO_NOT_USE
   flag for multi-integration data. [#60]